Refactor 1 - Split main into modules
